requirements.txt
PyJWT==1.7.1
secrets==1.0.2
